feat: reorder files in ptrace_binding, migrate personality to nanobind
