+++ libdebug/native/libdebug_linux_binding.pyi
@@ -1,0 +1,7 @@
+
+
+def disable_aslr() -> None:
+    """Disables ASLR for the current process."""
+
+def enable_aslr() -> None:
+    """Enables ASLR for the current process."""
